+++ src/sll/api/json_api.c
@@ -2,7 +2,6 @@
 #include <sll/api.h>
 #include <sll/api/json.h>
 #include <sll/array.h>
-#include <sll/assembly.h>
 #include <sll/common.h>
 #include <sll/file.h>
 #include <sll/handle.h>
@@ -13,6 +12,7 @@
 #include <sll/string.h>
 #include <sll/types.h>
 #include <sll/util.h>
+#include <sll/vm.h>
 #include <math.h>