+++ src/sll/api/sys_api.c
@@ -13,6 +13,7 @@
 #include <sll/types.h>
 #include <sll/util.h>
 #include <sll/version.h>
+#include <sll/vm.h>
 
 
 
@@ -139,10 +140,16 @@ __API_FUNC(sys_get_platform){
 
 
 __API_FUNC(sys_get_version){
-	sll_array_create(3,out);
-	out->v[0]=SLL_FROM_INT(SLL_VERSION_MAJOR);
-	out->v[1]=SLL_FROM_INT(SLL_VERSION_MINOR);
-	out->v[2]=SLL_FROM_INT(SLL_VERSION_PATCH);
+	sll_object_t* dt[3]={
+		SLL_FROM_INT(SLL_VERSION_MAJOR),
+		SLL_FROM_INT(SLL_VERSION_MINOR),
+		SLL_FROM_INT(SLL_VERSION_PATCH)
+	};
+	sll_object_t* o=sll_create_object_type(sll_current_runtime_data->tt,(sll_object_type_t)a,dt,3);
+	SLL_RELEASE(dt[0]);
+	SLL_RELEASE(dt[1]);
+	SLL_RELEASE(dt[2]);
+	return o;
 }