+++ src/sll/api/util_api.c
@@ -1,8 +1,8 @@
 #include <sll/_sll_internal.h>
 #include <sll/api.h>
-#include <sll/assembly.h>
 #include <sll/common.h>
 #include <sll/types.h>
+#include <sll/vm.h>