+++ src/sll/operator.c
@@ -2,7 +2,6 @@
 #include <sll/api/memory.h>
 #include <sll/api/string.h>
 #include <sll/array.h>
-#include <sll/assembly.h>
 #include <sll/common.h>
 #include <sll/gc.h>
 #include <sll/handle.h>
@@ -14,6 +13,7 @@
 #include <sll/string.h>
 #include <sll/types.h>
 #include <sll/util.h>
+#include <sll/vm.h>
 #include <math.h>